Add page view test for admin Add page add test (flow)

+++ src/main/java/ru/sspk/ssdmd/dao/QuestionDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.sspk.ssdmd.model.entity.Question;
 
 public interface QuestionDao extends JpaRepository<Question, Long> {
-    Question findById(int id);
+    Question findById(long id);
 
 }
